Make `resolver.rs` into a module.
With platform-specific modules and a separate unit-test module.
