+++ nym-vpn-core/crates/nym-vpn-lib/Cargo.toml
@@ -14,9 +14,18 @@ workspace = true
 [dependencies]
 anyhow.workspace = true
 async-trait.workspace = true
+base64.workspace = true
 bincode.workspace = true
-bip39.workspace = true
+bs58.workspace = true
 bytes.workspace = true
+
+ed25519-dalek = { workspace = true, features = [
+    "pkcs8",
+    "alloc",
+    "pem",
+    "rand_core",
+] }
+
 futures.workspace = true
 hex.workspace = true
 hickory-resolver = { workspace = true, features = [
@@ -30,47 +39,6 @@ http-body-util = { workspace = true }
 ipnetwork.workspace = true
 itertools.workspace = true
 log-panics.workspace = true
-opentelemetry.workspace = true
-opentelemetry_sdk.workspace = true
-rand.workspace = true
-reqwest = { workspace = true, features = ["socks", "rustls-tls"] }
-serial_test.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-semver = { workspace = true }
-sentry = { workspace = true, features = [
-    "backtrace",
-    "contexts",
-    "panic",
-    "reqwest",
-    "rustls",
-    "tracing",
-    "logs",
-] }
-strum.workspace = true
-toml.workspace = true
-thiserror.workspace = true
-time.workspace = true
-tokio = { workspace = true, features = [
-    "process",
-    "rt-multi-thread",
-    "fs",
-    "sync",
-] }
-tokio-stream.workspace = true
-tokio-util = { workspace = true, features = ["codec", "rt"] }
-tracing-appender.workspace = true
-tracing-subscriber = { workspace = true, features = [
-    "env-filter",
-    "ansi",
-    "json",
-] }
-tracing.workspace = true
-tracing-opentelemetry.workspace = true
-tun = { workspace = true, features = ["async"] }
-typed-builder.workspace = true
-url.workspace = true
-urlencoding.workspace = true
 
 nym-bandwidth-controller.workspace = true
 nym-bin-common.workspace = true
@@ -106,24 +74,59 @@ nym-vpn-store.workspace = true
 nym-wg-go = { workspace = true, features = ["amnezia"] }
 nym-wg-metadata-client.workspace = true
 
-ed25519-dalek = { workspace = true, features = [
-    "pkcs8",
-    "alloc",
-    "pem",
-    "rand_core",
-] }
+opentelemetry.workspace = true
+opentelemetry_sdk.workspace = true
 quinn.workspace = true
+quinn-proto.workspace = true
+rand.workspace = true
+reqwest = { workspace = true, features = ["socks", "rustls-tls"] }
+
 rustls = { workspace = true, default-features = false, features = [
     "std",
     "ring",
     "logging",
 ] }
-quinn-proto.workspace = true
+
+serial_test.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+semver = { workspace = true }
+sentry = { workspace = true, features = [
+    "backtrace",
+    "contexts",
+    "panic",
+    "reqwest",
+    "rustls",
+    "tracing",
+    "logs",
+] }
+strum.workspace = true
+toml.workspace = true
+thiserror.workspace = true
+time.workspace = true
+tokio = { workspace = true, features = [
+    "process",
+    "rt-multi-thread",
+    "fs",
+    "sync",
+] }
+tokio-stream.workspace = true
+tokio-util = { workspace = true, features = ["codec", "rt"] }
+tracing-appender.workspace = true
+tracing-subscriber = { workspace = true, features = [
+    "env-filter",
+    "ansi",
+    "json",
+] }
+tracing.workspace = true
+tracing-opentelemetry.workspace = true
+tun = { workspace = true, features = ["async"] }
+url.workspace = true
+urlencoding.workspace = true
 webpki-roots.workspace = true
 x509-parser = { workspace = true, features = ["verify"] }
-bs58.workspace = true
-base64.workspace = true
 
+# UNIX-specific dependencies
 [target.'cfg(unix)'.dependencies]
 nix = { workspace = true, features = [
     "socket",
@@ -133,36 +136,42 @@ nix = { workspace = true, features = [
     "ioctl",
 ] }
 
+# Windows-specific dependencies
 [target.'cfg(windows)'.dependencies]
-windows = { workspace = true, features = ["Win32_NetworkManagement_Ndis"] }
-winreg.workspace = true
 nym-windows.workspace = true
 nym-routing.workspace = true
 nym-dns.workspace = true
 nym-firewall.workspace = true
 nym-firewall-config.workspace = true
+socket2.workspace = true
+windows = { workspace = true, features = ["Win32_NetworkManagement_Ndis"] }
+winreg.workspace = true
 
-[target.'cfg(any(target_os = "linux", target_os = "macos"))'.dependencies]
+# Desktop-specific dependencies
+[target.'cfg(any(target_os = "linux", target_os = "macos", target_os = "windows"))'.dependencies]
 hickory-server = { workspace = true, features = ["resolver"] }
 nym-routing.workspace = true
 nym-dns.workspace = true
 nym-firewall.workspace = true
 nym-firewall-config.workspace = true
 
-[target.'cfg(target_os = "android")'.dependencies]
-tracing-android.workspace = true
-rand.workspace = true
-
+# Apple-specific dependencies
 [target.'cfg(any(target_os = "macos", target_os = "ios"))'.dependencies]
 tracing-oslog.workspace = true
 
+# Android-specific dependencies
+[target.'cfg(target_os = "android")'.dependencies]
+rand.workspace = true
+tracing-android.workspace = true
+
+# iOS-specific dependencies
 [target.'cfg(target_os = "ios")'.dependencies]
 debounced.workspace = true
 dispatch2.workspace = true
 
 [dev-dependencies]
-tempfile.workspace = true
 pretty_assertions.workspace = true
+tempfile.workspace = true
 
 [build-dependencies]
 vergen-gitcl = { workspace = true, default-features = false, features = [
